feat(plasma-temple): item - exports by platforms
