+++ packages/plasma-temple/src/components/Item/ItemBackground/ItemBackground.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import { DeviceKind } from '@sberdevices/plasma-ui';
 
-import { deviceFamily } from '../../../utils/deviceFamily';
+import { DeviceComponent } from '../../DeviceComponent/DeviceComponent';
 
-import { ItemBackgroundCommon, FullScreenBackgroundProps } from './ItemBackground@common';
+import { ItemBackgroundCommon, ItemBackgroundProps } from './ItemBackground@common';
 import { ItemBackgroundMobile } from './ItemBackground@mobile';
 
-export const mapDeviceToCartItem: Record<DeviceKind, React.FC<FullScreenBackgroundProps>> = {
-    sberBox: ItemBackgroundCommon,
-    sberPortal: ItemBackgroundCommon,
-    mobile: ItemBackgroundMobile,
-};
-
 /** Компонент, используемый для отображения фона на странице сущности */
-export const ItemBackground = mapDeviceToCartItem[deviceFamily];
+export const ItemBackground: React.FC<ItemBackgroundProps> = (props) => (
+    <DeviceComponent
+        sberbox={ItemBackgroundCommon}
+        sberportal={ItemBackgroundCommon}
+        mobile={ItemBackgroundMobile}
+        props={props}
+    />
+);
