+++ packages/plasma-temple/src/components/Item/ItemBackground/ItemBackground@common.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import { FullScreenBackground } from '../../FullScreenBackground/FullScreenBackground';
 
-export interface FullScreenBackgroundProps {
+export interface ItemBackgroundProps {
     src: string;
 }
 
@@ -11,6 +11,6 @@ const StyledFullScreenBackground = styled(FullScreenBackground)`
     z-index: -1;
 `;
 
-export const ItemBackgroundCommon: React.FC<FullScreenBackgroundProps> = ({ src }) => (
+export const ItemBackgroundCommon: React.FC<ItemBackgroundProps> = ({ src }) => (
     <StyledFullScreenBackground src={src} />
 );
