+++ packages/plasma-temple/src/components/Item/ItemEntities/ItemEntities.tsx
@@ -1,14 +1,16 @@
 import React from 'react';
-import { detectDevice, DeviceKind } from '@sberdevices/plasma-ui/utils';
+
+import { DeviceComponent } from '../../DeviceComponent/DeviceComponent';
 
 import { ItemEntitiesCommon, ItemEntitiesProps } from './ItemEntities@common';
 import { ItemEntitiesMobile } from './ItemEntities@mobile';
 
-const mapDeviceToItemEntities: Record<DeviceKind, React.FC<ItemEntitiesProps>> = {
-    sberBox: ItemEntitiesCommon,
-    sberPortal: ItemEntitiesCommon,
-    mobile: ItemEntitiesMobile,
-};
-
 /** Компонент, используемы для отображения списка карточек */
-export const ItemEntities = React.memo(mapDeviceToItemEntities[detectDevice()]) as typeof ItemEntitiesCommon;
+export const ItemEntities = React.memo<ItemEntitiesProps>((props) => (
+    <DeviceComponent
+        sberbox={ItemEntitiesCommon}
+        sberportal={ItemEntitiesCommon}
+        mobile={ItemEntitiesMobile}
+        props={props}
+    />
+)) as typeof ItemEntitiesCommon;
