+++ packages/plasma-temple/src/components/Item/index@mobile.ts
@@ -1,0 +1,4 @@
+export { ItemMobile as Item } from './Item@mobile';
+export { ItemDetails } from './ItemDetails/ItemDetails';
+export { ItemMainSection } from './ItemMainSection/ItemMainSection';
+export { ItemEntitiesMobile as ItemEntities } from './ItemEntities/ItemEntities@mobile';
