+++ packages/plasma-temple/src/components/Item/index@sberbox.ts
@@ -1,0 +1,4 @@
+export { ItemSberBox as Item } from './Item@sberbox';
+export { ItemDetails } from './ItemDetails/ItemDetails';
+export { ItemMainSection } from './ItemMainSection/ItemMainSection';
+export { ItemEntitiesCommon as ItemEntities } from './ItemEntities/ItemEntities@common';
