+++ packages/plasma-temple/src/components/Item/index@sberportal.ts
@@ -1,0 +1,4 @@
+export { ItemSberPortal as Item } from './Item@sberportal';
+export { ItemDetails } from './ItemDetails/ItemDetails';
+export { ItemMainSection } from './ItemMainSection/ItemMainSection';
+export { ItemEntitiesCommon as ItemEntities } from './ItemEntities/ItemEntities@common';
