+++ packages/plasma-temple/src/components/index@mobile.ts
@@ -1,1 +1,2 @@
 export { StateLayoutMobile as StateLayout } from './StateLayout/StateLayout@mobile';
+export * from './Item/index@mobile';
