+++ packages/plasma-temple/src/components/index@sberbox.ts
@@ -1,1 +1,2 @@
 export { StateLayoutSberBox as StateLayout } from './StateLayout/StateLayout@sberbox';
+export * from './Item/index@sberbox';
