+++ packages/plasma-temple/src/components/index@sberportal.ts
@@ -1,1 +1,2 @@
 export { StateLayoutSberPortal as StateLayout } from './StateLayout/StateLayout@sberportal';
+export * from './Item/index@sberportal';
